requirements.txt
pillow
pypdf2
ttkthemes
docx
pillow